Split page routes into server/client components, route SignalR through proxy
- Extract "use client" logic from 6 page.tsx files into page.client.tsx
  siblings so Next.js can keep the route entry point as a Server Component
- Dynamic routes (messages/[chatId], tasks/[id]) resolve params on the
  server and pass the resolved value as a prop to the client component
- Remove NEXT_PUBLIC_BACKEND_URL from chat-hub.ts; SignalR now connects
  to relative /hubs/chat so all traffic goes through the Next.js proxy
- Add /hubs/:path* rewrite in next.config.ts forwarding to API_URL,
  keeping the backend origin in a single server-side env var
